Enemy Damage
Enemies now damage the player through collision. This needs to be ironed out, but it works similarly to projectiles where the player hitbox detects bodies that have entered the Area2D and apply their damage to the player. Not sure how this will work with multiple enemies damaging and the same time yet