Enemy Damage Implemented
Enemy and player hitboxes have been fixed so that the enemy will damage the player while in range of attack. This damage does not scale with multiple enemies attacking yet, because the damage calculation is done in the player class at the moment. The next step will be to have each enemy instance handle their own damage to the player, as well as making a death state