Death State and Game Over Screen
Implemented death animation, game over sequence, and buttons to either restart or quit the game. Also made a gamemanager script on the World node that will be used to manage the overarching game systems (including UI and the various game state classes). The restart function is currently a bit broken because the player still can't move afterwards but I'll fix that later LOL